Simplify lint/format functions (API BREAKING) * The functions do not accept the Params class as parameter anymore but only the ExperimentalParams. * The VisitorProvider parameter is no longer accepted. It will be created internally based on information of the ExperimentParams.
Previously the lint/format function required a VisitorProvider as parameter. This parameter had to be instantiated early in the call hierarchy, so that the same VisitorProvider instance could be reused for processing all files. This is important as during initialization of the VisitorProvider, the order in which the rules are executed, is determined. While doing so, this rule order is also logged. Late initialization of the VisitorProvider would result initializing the VisitorProvider for each file which also would result in a performance loss and lots of duplicated log lines.

Now late initialization is possible without causing the performance loss nor the superfluous logging. The rule execution order is now determined by the RuleSorter which is extracted from the VisitorProvider. The RuleSorter only determines the rule execution order whenever a new unique set of rule ids is provided. Other a cached instance is returned. As a result the logging is only printed for unique combinations of rules. Also, the QualifiedRuleId has been extracted from the VisitorProvider.

The VisitorProvider has been changed to an internal class.
