+++ ktlint-core/src/main/kotlin/com/pinterest/ktlint/core/KtLint.kt
@@ -12,6 +12,7 @@ import com.pinterest.ktlint.core.internal.EditorConfigLoader.Companion.convertTo
 import com.pinterest.ktlint.core.internal.KotlinPsiFileFactoryProvider
 import com.pinterest.ktlint.core.internal.LineAndColumn
 import com.pinterest.ktlint.core.internal.SuppressionLocator
+import com.pinterest.ktlint.core.internal.VisitorProvider
 import com.pinterest.ktlint.core.internal.buildPositionInTextLocator
 import com.pinterest.ktlint.core.internal.buildSuppressedRegionsLocator
 import com.pinterest.ktlint.core.internal.noSuppression
@@ -44,31 +45,6 @@ public object KtLint {
 
     @OptIn(FeatureInAlphaState::class)
     private val editorConfigGenerator = EditorConfigGenerator(editorConfigLoader)
-
-    /**
-     * @param fileName path of file to lint/format
-     * @param text Contents of file to lint/format
-     * @param ruleSets a collection of "RuleSet"s used to validate source
-     * @param userData Map of user options
-     * @param cb callback invoked for each lint error
-     * @param script true if this is a Kotlin script file
-     * @param editorConfigPath optional path of the .editorconfig file (otherwise will use working directory)
-     * @param debug True if invoked with the --debug flag
-     */
-    @Deprecated(
-        message = "Marked for removal in Ktlint 0.46.",
-        replaceWith = ReplaceWith("ExperimentalParams")
-    )
-    public data class Params(
-        val fileName: String? = null,
-        val text: String,
-        val ruleSets: Iterable<RuleSet>,
-        val userData: Map<String, String> = emptyMap(),
-        val cb: (e: LintError, corrected: Boolean) -> Unit,
-        val script: Boolean = false,
-        val editorConfigPath: String? = null,
-        val debug: Boolean = false
-    )
 
     /**
      * @param fileName path of file to lint/format
@@ -145,64 +121,18 @@ public object KtLint {
             .toSet()
     }
 
-    @Deprecated(
-        message = "Marked for removal in Ktlint 0.46.",
-        replaceWith = ReplaceWith("ExperimentalParams")
-    )
-    @OptIn(FeatureInAlphaState::class)
-    private fun toExperimentalParams(params: Params): ExperimentalParams =
-        ExperimentalParams(
-            fileName = params.fileName,
-            text = params.text,
-            ruleSets = params.ruleSets,
-            userData = params.userData,
-            cb = params.cb,
-            script = params.script,
-            editorConfigPath = params.editorConfigPath,
-            debug = params.debug
-        )
-
     /**
      * Check source for lint errors.
      *
      * @throws ParseException if text is not a valid Kotlin code
      * @throws RuleExecutionException in case of internal failure caused by a bug in rule implementation
      */
-    @Deprecated(
-        message = "Marked for removal in Ktlint 0.46. Move '.editorconfig' properties from 'Params.userData' to 'ExperimentalParam.editorConfigOverride'.",
-        replaceWith = ReplaceWith("lint(toExperimentalParams(params))")
-    )
-    @OptIn(FeatureInAlphaState::class)
-    public fun lint(params: Params) =
-        lint(toExperimentalParams(params))
-
-    @OptIn(FeatureInAlphaState::class)
-    public fun lint(experimentalParams: ExperimentalParams) {
-        lint(
-            experimentalParams,
-            VisitorProvider(
-                ruleSets = experimentalParams.ruleSets,
-                debug = experimentalParams.debug
-            )
-        )
-    }
-
-    /**
-     * Check source for lint errors.
-     *
-     * @throws ParseException if text is not a valid Kotlin code
-     * @throws RuleExecutionException in case of internal failure caused by a bug in rule implementation
-     */
-    @FeatureInAlphaState
-    public fun lint(
-        params: ExperimentalParams,
-        visitorProvider: VisitorProvider
-    ) {
+    public fun lint(params: ExperimentalParams) {
         val psiFileFactory = kotlinPsiFileFactoryProvider.getKotlinPsiFileFactory(params.isInvokedFromCli)
         val preparedCode = prepareCodeForLinting(psiFileFactory, params)
         val errors = mutableListOf<LintError>()
 
-        visitorProvider
+        VisitorProvider(params.ruleSets, params.debug)
             .visitor(params.ruleSets, preparedCode.rootNode)
             .invoke { node, rule, fqRuleId ->
                 // fixme: enforcing suppression based on node.startOffset is wrong
@@ -333,59 +263,17 @@ public object KtLint {
      * @throws ParseException if text is not a valid Kotlin code
      * @throws RuleExecutionException in case of internal failure caused by a bug in rule implementation
      */
-    @Deprecated(
-        message = "Marked for removal in Ktlint 0.46. Move '.editorconfig' properties from 'Params.userData' to 'ExperimentalParam.editorConfigOverride'.",
-        replaceWith = ReplaceWith("lint(toExperimentalParams(params))")
-    )
-    @OptIn(FeatureInAlphaState::class)
-    public fun format(params: Params): String =
-        format(toExperimentalParams(params))
-
-    @OptIn(FeatureInAlphaState::class)
-    public fun format(experimentalParams: ExperimentalParams): String {
-        return format(
-            experimentalParams,
-            experimentalParams.ruleSets,
-            VisitorProvider(
-                ruleSets = experimentalParams.ruleSets,
-                debug = experimentalParams.debug
-            )
-        )
-    }
-
-    /**
-     * Fix style violations.
-     *
-     * @throws ParseException if text is not a valid Kotlin code
-     * @throws RuleExecutionException in case of internal failure caused by a bug in rule implementation
-     */
-    @Deprecated(
-        message = "Marked for removal in Ktlint 0.46. Overrides of '.editorconfig' properties no longer should be " +
-            "passed via the 'Params.userData' but via the 'ExperimentalParam.editorConfigOverride' parameter. The " +
-            "ruleSets have to be provided via the 'ExperimentalParams.ruleSets' parameter.",
-        replaceWith = ReplaceWith("format(params, visitorProvider)")
-    )
-    @FeatureInAlphaState
-    public fun format(
-        params: ExperimentalParams,
-        ruleSets: Iterable<RuleSet>,
-        visitorProvider: VisitorProvider
-    ): String =
-        format(
-            params.copy(ruleSets = ruleSets),
-            visitorProvider
-        )
-
-    public fun format(
-        params: ExperimentalParams,
-        visitorProvider: VisitorProvider
-    ): String {
+    public fun format(params: ExperimentalParams): String {
         val hasUTF8BOM = params.text.startsWith(UTF8_BOM)
         val psiFileFactory = kotlinPsiFileFactoryProvider.getKotlinPsiFileFactory(params.isInvokedFromCli)
         val preparedCode = prepareCodeForLinting(psiFileFactory, params)
 
         var tripped = false
         var mutated = false
+        val visitorProvider = VisitorProvider(
+            ruleSets = params.ruleSets,
+            debug = params.debug
+        )
         visitorProvider
             .visitor(
                 params.ruleSets,
