+++ ktlint-core/src/main/kotlin/com/pinterest/ktlint/core/internal/RuleSorter.kt
@@ -1,165 +1,54 @@
-package com.pinterest.ktlint.core
+package com.pinterest.ktlint.core.internal
 
-import com.pinterest.ktlint.core.ast.visit
+import com.pinterest.ktlint.core.Rule
+import com.pinterest.ktlint.core.RuleSet
 import kotlin.reflect.KClass
-import org.jetbrains.kotlin.com.intellij.lang.ASTNode
 
-public class VisitorProvider(
-    ruleSets: Iterable<RuleSet>,
-    private val debug: Boolean
-) {
-    private val ruleReferences: List<RuleReference> =
-        VisitorProviderInitializer(ruleSets, debug).getRulePreferences()
+/**
+ * Normally, the [RuleSorter] should be used as a singleton as it logs the order in which the rules are executed. Only
+ * in specific unit tests scenario's it might be needed to recreate the class when running unit tests for the same set
+ * of rule ids but having different [Rule.VisitorModifier]s.
+ */
+internal class RuleSorter {
+    private val ruleSetsToRuleReferencesMap = mutableMapOf<Int, List<RuleReference>>()
 
-    internal fun visitor(
+    @Synchronized
+    fun getSortedRules(
         ruleSets: Iterable<RuleSet>,
-        rootNode: ASTNode,
-        concurrent: Boolean = true
-    ): ((node: ASTNode, rule: Rule, fqRuleId: String) -> Unit) -> Unit {
-        val enabledRuleReferences =
-            ruleReferences
-                .filter { ruleReference -> isNotDisabled(rootNode, ruleReference.toQualifiedRuleId()) }
-        val enabledQualifiedRuleIds = enabledRuleReferences.map { it.toQualifiedRuleId() }
-        val enabledRules = ruleSets
-            .flatMap { ruleSet ->
+        debug: Boolean
+    ): List<RuleReference> {
+        val id = ruleSets
+            .asSequence()
+            .map { ruleSet ->
                 ruleSet
                     .rules
-                    .filter { rule -> toQualifiedRuleId(ruleSet.id, rule.id) in enabledQualifiedRuleIds }
-                    .filter { rule -> isNotDisabled(rootNode, toQualifiedRuleId(ruleSet.id, rule.id)) }
-                    .map { rule -> "${toQualifiedRuleId(ruleSet.id, rule.id)}" to rule }
-            }.toMap()
-        if (debug && enabledRules.isEmpty()) {
-            println(
-                "[DEBUG] Skipping file as no enabled rules are found to be executed"
-            )
-            return { _ -> }
-        }
-        val ruleReferencesToBeSkipped =
-            ruleReferences
-                .filter { ruleReference ->
-                    ruleReference.runAfterRule != null &&
-                        ruleReference.runAfterRule.runOnlyWhenOtherRuleIsEnabled &&
-                        enabledRules[ruleReference.runAfterRule.ruleId.toQualifiedRuleId()] == null
-                }
-        if (debug && ruleReferencesToBeSkipped.isNotEmpty()) {
-            ruleReferencesToBeSkipped
-                .forEach {
-                    println(
-                        "[DEBUG] Skipping rule with id '${it.toQualifiedRuleId()}'. This rule has to run after rule with " +
-                            "id '${it.runAfterRule?.ruleId?.toQualifiedRuleId()}' and will not run in case that rule is " +
-                            "disabled."
-                    )
-                }
-        }
-        val ruleReferenceWithoutEntriesToBeSkipped = enabledRuleReferences - ruleReferencesToBeSkipped
-        if (debug && ruleReferenceWithoutEntriesToBeSkipped.isEmpty()) {
-            println(
-                "[DEBUG] Skipping file as no enabled rules are found to be executed"
-            )
-            return { _ -> }
-        }
-        return if (concurrent) {
-            concurrentVisitor(enabledRules, ruleReferenceWithoutEntriesToBeSkipped, rootNode)
-        } else {
-            sequentialVisitor(enabledRules, ruleReferenceWithoutEntriesToBeSkipped, rootNode)
-        }
-    }
-
-    private fun concurrentVisitor(
-        enabledRules: Map<String, Rule>,
-        ruleReferences: List<RuleReference>,
-        rootNode: ASTNode
-    ): ((node: ASTNode, rule: Rule, fqRuleId: String) -> Unit) -> Unit {
-        return { visit ->
-            rootNode.visit { node ->
-                ruleReferences
-                    .forEach { ruleReference ->
-                        if (node == rootNode || !ruleReference.runOnRootNodeOnly) {
-                            enabledRules[ruleReference.toQualifiedRuleId()]
-                                ?.let { rule ->
-                                    visit(node, rule, ruleReference.toShortenedQualifiedRuleId())
+                    .map { rule ->
+                        toQualifiedRuleId(ruleSet.id, rule.id)
+                    }
+            }.flatten()
+            .sorted()
+            .joinToString(prefix = "rule-ids=[", separator = ",", postfix = "]")
+            .plus(",debug=$debug")
+        val ruleSetsHashCode = id.hashCode()
+        return ruleSetsToRuleReferencesMap
+            // If the VisitorProvider has been created before for the same list of RuleSets than reuse the list of
+            // ruleReferences as computation is relative expensive but more important because it generates as lot of
+            // logging. Only during unit tests it is normal behavior to have a frequent changing list of rule sets.
+            .getOrPut(ruleSetsHashCode) {
+                ruleSets
+                    .flatMap { it.toRuleReferences() }
+                    .sortedWith(defaultRuleExecutionOrderComparator())
+                    .applyRunAfterRuleToRuleExecutionOrder(debug)
+                    .also { ruleReferences ->
+                        if (debug) {
+                            ruleReferences
+                                .map { toQualifiedRuleId(it.ruleSetId, it.ruleId) }
+                                .joinToString(prefix = "[DEBUG] Rules will be executed in order below (unless disabled):") {
+                                    "\n           - $it"
                                 }
+                                .let { println(it) }
                         }
                     }
-            }
-        }
-    }
-
-    private fun sequentialVisitor(
-        enabledRules: Map<String, Rule>,
-        ruleReferences: List<RuleReference>,
-        rootNode: ASTNode
-    ): ((node: ASTNode, rule: Rule, fqRuleId: String) -> Unit) -> Unit {
-        return { visit ->
-            ruleReferences
-                .forEach { ruleReference ->
-                    enabledRules[ruleReference.toQualifiedRuleId()]
-                        ?.let { rule ->
-                            if (ruleReference.runOnRootNodeOnly) {
-                                visit(rootNode, rule, ruleReference.toShortenedQualifiedRuleId())
-                            } else {
-                                rootNode.visit { node -> visit(node, rule, ruleReference.toShortenedQualifiedRuleId()) }
-                            }
-                        }
-                }
-        }
-    }
-
-    private fun RuleReference.toShortenedQualifiedRuleId() =
-        if (ruleSetId == "standard") {
-            ruleId
-        } else {
-            toQualifiedRuleId()
-        }
-
-    private fun isNotDisabled(rootNode: ASTNode, qualifiedRuleId: String): Boolean =
-        rootNode
-            .getUserData(KtLint.DISABLED_RULES)
-            .orEmpty()
-            .none {
-                // The rule set id in the disabled_rules setting may be omitted for rules in the standard rule set
-                it.toQualifiedRuleId() == qualifiedRuleId
-            }
-}
-
-private fun RuleReference.toQualifiedRuleId() =
-    toQualifiedRuleId(ruleSetId, ruleId)
-
-private fun toQualifiedRuleId(
-    ruleSetId: String,
-    ruleId: String
-) =
-    if (ruleId.startsWith("$ruleSetId:")) {
-        ruleId
-    } else {
-        "$ruleSetId:$ruleId"
-    }
-
-private fun String.toQualifiedRuleId() =
-    if (contains(":")) {
-        this
-    } else {
-        "standard:$this"
-    }
-
-private class VisitorProviderInitializer(
-    val ruleSets: Iterable<RuleSet>,
-    val debug: Boolean
-) {
-    fun getRulePreferences(): List<RuleReference> {
-        return ruleSets
-            .flatMap { it.toRuleReferences() }
-            .sortedWith(defaultRuleExecutionOrderComparator)
-            .applyRunAfterRuleToRuleExecutionOrder()
-            .also { ruleReferences ->
-                if (debug) {
-                    ruleReferences
-                        .map { toQualifiedRuleId(it.ruleSetId, it.ruleId) }
-                        .joinToString(prefix = "[DEBUG] Rules will be executed in order below (unless disabled):") {
-                            "\n           - $it"
-                        }
-                        .let { println(it) }
-                }
             }
     }
 
@@ -241,7 +130,30 @@ private class VisitorProviderInitializer(
         )
     }
 
-    private fun List<RuleReference>.applyRunAfterRuleToRuleExecutionOrder(): List<RuleReference> {
+    private fun defaultRuleExecutionOrderComparator() =
+        // The sort order below should guarantee a stable order of the rule between multiple invocations of KtLint given
+        // the same set of input parameters. There should be no dependency on data ordering outside this class.
+        compareBy<RuleReference> {
+            if (it.runAsLateAsPossible) {
+                1
+            } else {
+                0
+            }
+        }.thenBy {
+            if (it.runOnRootNodeOnly) {
+                0
+            } else {
+                1
+            }
+        }.thenBy {
+            when (it.ruleSetId) {
+                "standard" -> 0
+                "experimental" -> 1
+                else -> 2
+            }
+        }.thenBy { it.ruleId }
+
+    private fun List<RuleReference>.applyRunAfterRuleToRuleExecutionOrder(debug: Boolean): List<RuleReference> {
         // The new list of rule references retains the order of the original list of rule references as much as
         // possible. Rule references will only be deferred till later in the list when needed.
         val newRuleReferences = mutableListOf<RuleReference>()
@@ -332,43 +244,4 @@ private class VisitorProviderInitializer(
 
     private fun RuleReference.runsAfter(ruleReference: RuleReference) =
         ruleReference.runAfterRule?.ruleId == toQualifiedRuleId()
-
-    private companion object {
-        val defaultRuleExecutionOrderComparator =
-            // The sort order below should guarantee a stable order of the rule between multiple invocations of KtLint given
-            // the same set of input parameters. There should be no dependency on data ordering outside this class.
-            compareBy<RuleReference> {
-                if (it.runAsLateAsPossible) {
-                    1
-                } else {
-                    0
-                }
-            }.thenBy {
-                if (it.runOnRootNodeOnly) {
-                    0
-                } else {
-                    1
-                }
-            }.thenBy {
-                when (it.ruleSetId) {
-                    "standard" -> 0
-                    "experimental" -> 1
-                    else -> 2
-                }
-            }.thenBy { it.ruleId }
-    }
 }
-
-private data class RuleReference(
-    val ruleId: String,
-    val ruleSetId: String,
-    val runOnRootNodeOnly: Boolean,
-    val runAsLateAsPossible: Boolean,
-    val runAfterRule: Rule.VisitorModifier.RunAfterRule?
-)
-
-private data class RunAfter(
-    val ruleId: String,
-    val loadOnlyWhenOtherRuleIsLoaded: Boolean,
-    val runOnlyWhenOtherRuleIsEnabled: Boolean
-)
