+++ ktlint/src/main/kotlin/com/pinterest/ktlint/Main.kt
@@ -12,7 +12,6 @@ import com.pinterest.ktlint.core.ReporterProvider
 import com.pinterest.ktlint.core.RuleExecutionException
 import com.pinterest.ktlint.core.RuleSet
 import com.pinterest.ktlint.core.RuleSetProvider
-import com.pinterest.ktlint.core.VisitorProvider
 import com.pinterest.ktlint.core.initKtLintKLogger
 import com.pinterest.ktlint.core.internal.containsLintError
 import com.pinterest.ktlint.core.internal.loadBaseline
@@ -277,13 +276,11 @@ class KtlintCommandLine {
 
         reporter.beforeAll()
         val ruleSets = ruleSetProviders.map { it.value.get() }
-        val visitorProvider = VisitorProvider(ruleSets, debug)
         if (stdin) {
-            lintStdin(ruleSetProviders, visitorProvider, userData, reporter)
+            lintStdin(ruleSetProviders, userData, reporter)
         } else {
             lintFiles(
                 ruleSetProviders,
-                visitorProvider,
                 userData,
                 baselineResults.baselineRules,
                 reporter
@@ -310,7 +307,6 @@ class KtlintCommandLine {
 
     private fun lintFiles(
         ruleSetProviders: Map<String, RuleSetProvider>,
-        visitorProvider: VisitorProvider,
         userData: Map<String, String>,
         baseline: Map<String, List<LintError>>?,
         reporter: Reporter
@@ -327,8 +323,7 @@ class KtlintCommandLine {
                         file.readText(),
                         userData,
                         baseline?.get(file.relativeRoute),
-                        ruleSets,
-                        visitorProvider
+                        ruleSets
                     )
                 }
             }
@@ -337,7 +332,6 @@ class KtlintCommandLine {
 
     private fun lintStdin(
         ruleSetProviders: Map<String, RuleSetProvider>,
-        visitorProvider: VisitorProvider,
         userData: Map<String, String>,
         reporter: Reporter
     ) {
@@ -348,8 +342,7 @@ class KtlintCommandLine {
                 String(System.`in`.readBytes()),
                 userData,
                 null,
-                ruleSetProviders.map { it.value.get() },
-                visitorProvider
+                ruleSetProviders.map { it.value.get() }
             ),
             reporter
         )
@@ -397,8 +390,7 @@ class KtlintCommandLine {
         fileContent: String,
         userData: Map<String, String>,
         baselineErrors: List<LintError>?,
-        ruleSets: Iterable<RuleSet>,
-        visitorProvider: VisitorProvider
+        ruleSets: Iterable<RuleSet>
     ): List<LintErrorWithCorrectionInfo> {
         logger.trace {
             val fileLocation = if (fileName != KtLint.STDIN_FILE) File(fileName).location(relative) else fileName
@@ -413,17 +405,15 @@ class KtlintCommandLine {
                     ruleSets,
                     userData,
                     editorConfigPath,
-                    debug,
-                    { err, corrected ->
-                        if (!corrected) {
-                            if (baselineErrors == null || !baselineErrors.containsLintError(err)) {
-                                result.add(LintErrorWithCorrectionInfo(err, corrected))
-                                tripped.set(true)
-                            }
+                    debug
+                ) { err, corrected ->
+                    if (!corrected) {
+                        if (baselineErrors == null || !baselineErrors.containsLintError(err)) {
+                            result.add(LintErrorWithCorrectionInfo(err, corrected))
+                            tripped.set(true)
                         }
-                    },
-                    visitorProvider
-                )
+                    }
+                }
             } catch (e: Exception) {
                 result.add(LintErrorWithCorrectionInfo(e.toLintError(), false))
                 tripped.set(true)
@@ -442,7 +432,6 @@ class KtlintCommandLine {
                     fileName,
                     fileContent,
                     ruleSets,
-                    visitorProvider,
                     userData,
                     editorConfigPath,
                     debug
